Extract transcription IPC handlers to src/ipc/
main.ts is 2,476 lines with 46 IPC handlers across mixed domains. Splitting
each cluster into its own module lets us reason about (and test) one slice
at a time without dragging the rest of the file into scope.

This change moves transcribe-audio and cancel-transcription -- the entire
in-flight transcription cluster, including the activeTranscriptions
AbortController map and the renameAudioFile helper that's only used by the
post-transcription "Untitled_Meeting" rename path. main keeps owning the
singletons (geminiService, configService, the mainWindow ref) and exposes
them through a new IpcContext object so the moved handlers always see the
live references even after main re-assigns the underlying lets.

IpcContext also routes shared helpers (formatAiCredentialsError,
serializeTranscriptionError, sanitizeLiveNotes) and the maybeAutoSync
trigger via ctx, so the transcription module never imports back from
main.ts. The sister PRs (meetings-mgmt, google-drive) create the same
src/ipc/types.ts and src/ipc/register.ts scaffolding; whichever lands
first wins and the others rebase to union their ctx fields and register
calls into the merged scaffolding.

Behavior is unchanged -- this is a pure file move plus a thin DI seam.
