+++ src/ipc/register.ts
@@ -1,4 +1,5 @@
 import * as meetingsManagementIpc from './meetingsManagement';
+import * as transcriptionIpc from './transcription';
 import type { IpcContext } from './types';
 
 // Single wiring point for every src/ipc/<domain>.ts module that has been
@@ -9,4 +10,5 @@ import type { IpcContext } from './types';
 // stays touched only when the context surface changes.
 export function registerAllIpc(ctx: IpcContext): void {
   meetingsManagementIpc.register(ctx);
+  transcriptionIpc.register(ctx);
 }
